+++ zarrs_metadata_ext/src/codec/zarrs/squeeze.rs
@@ -8,14 +8,14 @@ use zarrs_metadata::ConfigurationSerialize;
 #[non_exhaustive]
 #[serde(untagged)]
 pub enum SqueezeCodecConfiguration {
-    /// Version 1.0.
-    V1(SqueezeCodecConfigurationV1),
+    /// Version 0.0 draft.
+    V0(SqueezeCodecConfigurationV0),
 }
 
 impl ConfigurationSerialize for SqueezeCodecConfiguration {}
 
-/// `squeeze` codec configuration parameters (version 1.0).
+/// `squeeze` codec configuration parameters (version 0.0 draft).
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq, Debug, Display)]
 #[serde(deny_unknown_fields)]
 #[display("{}", serde_json::to_string(self).unwrap_or_default())]
-pub struct SqueezeCodecConfigurationV1 {}
+pub struct SqueezeCodecConfigurationV0 {}