+++ zarrs_metadata_ext/CHANGELOG.md
@@ -13,5 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - rename `ArrayMetadataV2ToV3ConversionError` to `ArrayMetadataV2ToV3Error`
 - rename `InvalidPermutationError` to `TransposeOrderError`
+- change the suffix of experimental codec configurations from V1 to V0 (`gdeflate`, `squeeze`, `vlen`, `vlen_v2`)
 
 [unreleased]: https://github.com/zarrs/zarrs/compare/zarrs_metadata-v0.1.0...HEAD